DD-75 refactor : 보낸 꿀도장을 그룹별로 조회하는 GetSentStampByGroupService 내부 로직을 StampByGroupMapper로 변경
